+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- Re-added `insightvm_vulnerability` and `insightvm_finding_is_vulnerability`.
+
 ## 0.2.6 - 2021-02-12
 
 ### Fixed